bitset

+++ ip-address-filters/bitset_test.go
@@ -32,7 +32,7 @@ func BenchmarkBitset(b *testing.B) {
 
 	b.Run("sequential int", func(b *testing.B) {
 		for i := 0; i < b.N; i++ {
-			bit := uint(i)
+			bit := uint(uint32(i))
 			if bs.Test(bit) {
 				checkMatch(uint2ip(bit))
 			}
